+++ examples/equipment_unit/__init__.py
@@ -1,0 +1,3 @@
+from equipment_unit._api_client import EquipmentUnitClient
+
+__all__ = ["EquipmentUnitClient"]
