--- examples/equipment_unit/client/__init__.py
@@ -1,3 +1,0 @@
-from equipment_unit.client._api_client import EquipmentUnitClient
-
-__all__ = ["EquipmentUnitClient"]
