+++ examples/equipment_unit/data_classes/_equipment_module.py
