+++ examples/omni_multi/_api_client.py
@@ -128,8 +128,8 @@ class OmniMultiClient:
 
     Generated with:
         pygen = 0.99.15
-        cognite-sdk = 7.27.2
-        pydantic = 2.6.3
+        cognite-sdk = 7.30.1
+        pydantic = 2.6.4
 
     """
