+++ examples/scenario_instance/client/_api_client.py
@@ -20,8 +20,8 @@ class ScenarioInstanceClient:
 
     Generated with:
         pygen = 0.99.15
-        cognite-sdk = 7.27.2
-        pydantic = 2.6.3
+        cognite-sdk = 7.30.1
+        pydantic = 2.6.4
 
     Data Model:
         space: IntegrationTestsImmutable
